Avoid mentioning handshake in linkability text
The handshake itself is not particularly special in terms of
correlating payloads that belong to the same connection -- it is
only special in that it indicates the start of a connection.
So, don't mention it when discussing linkability, to avoid giving
the reader the impression that the handshake needs to be observed
in order for linking to be possible.